# openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoStorm",
    "description": "Observability for AI-driven software discovery",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get Alerts",
        "operationId": "listAlerts",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlertSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "acknowledged",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acknowledged"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AlertResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{alert_id}/acknowledge": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "summary": "Post Acknowledge Alert",
        "operationId": "acknowledgeAlert",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Acknowledgealert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/config": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get Alert Config",
        "operationId": "getAlertConfig",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertConfigResponse"
                  },
                  "title": "Response Getalertconfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Alerts"
        ],
        "summary": "Update Alert Config",
        "operationId": "updateAlertConfig",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertConfigResponse"
                  },
                  "title": "Response Updatealertconfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertConfigItem": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "alert_types": {
            "items": {
              "$ref": "#/components/schemas/AlertType"
            },
            "type": "array",
            "title": "Alert Types"
          },
          "min_severity": {
            "$ref": "#/components/schemas/AlertSeverity",
            "default": "info"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "channel",
          "endpoint"
        ],
        "title": "AlertConfigItem",
        "description": "Alert configuration item for request/response."
      },
      "AlertConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "alert_types": {
            "items": {
              "$ref": "#/components/schemas/AlertType"
            },
            "type": "array",
            "title": "Alert Types"
          },
          "min_severity": {
            "$ref": "#/components/schemas/AlertSeverity",
            "default": "info"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "channel",
          "endpoint",
          "created_at",
          "updated_at"
        ],
        "title": "AlertConfigResponse",
        "description": "Alert configuration detail."
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "alert_type": {
            "$ref": "#/components/schemas/AlertType"
          },
          "severity": {
            "$ref": "#/components/schemas/AlertSeverity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "is_acknowledged": {
            "type": "boolean",
            "title": "Is Acknowledged",
            "default": false
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "acknowledged_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "alert_type",
          "severity",
          "title",
          "message",
          "created_at"
        ],
        "title": "AlertResponse",
        "description": "Alert detail."
      },
      "AlertSeverity": {
        "type": "string",
        "enum": [
          "info",
          "warning",
          "critical"
        ],
        "title": "AlertSeverity",
        "description": "Severity levels for alerts."
      },
      "AlertType": {
        "type": "string",
        "enum": [
          "competitor_emergence",
          "configuration_error",
          "disappearance",
          "recommendation_share_drop",
          "position_degradation",
          "model_divergence",
          "citation_domain_shift"
        ],
        "title": "AlertType",
        "description": "Types of alerts the system can generate."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_AlertResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[AlertResponse]"
      },
...
      "UpdateAlertConfigRequest": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/AlertConfigItem"
            },
            "type": "array",
            "title": "Configs"
          }
        },
        "type": "object",
        "required": [
          "configs"
        ],
        "title": "UpdateAlertConfigRequest",
        "description": "Request to update alert configs."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}